+++ tests/poly1305-256-test.c
@@ -51,7 +51,7 @@ main()
   }
 
   bool ok = true;
-  for (int i = 0; i < sizeof(vectors) / sizeof(poly1305_test_vector); ++i) {
+  for (size_t i = 0; i < sizeof(vectors) / sizeof(poly1305_test_vector); ++i) {
     ok &= print_test(
       vectors[i].input_len, vectors[i].input, vectors[i].key, vectors[i].tag);
   }
